+++ packages/orq-rc/src/models/operations/createidentity.ts
@@ -139,7 +139,7 @@ export const CreateIdentityResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-05-15T03:37:24.790Z",
+    "2026-05-15T11:15:21.069Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {