+++ packages/orq-rc/RELEASES.md
@@ -9219,3 +9219,13 @@ Based on:
 - [typescript v4.10.0-rc.7] packages/orq-rc
 ### Releases
 - [NPM v4.10.0-rc.7] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.7 - packages/orq-rc
+
+## 2026-05-15 11:18:13
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.763.1 (2.884.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.10.0-rc.8] packages/orq-rc
+### Releases
+- [NPM v4.10.0-rc.8] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.8 - packages/orq-rc